requirements.txt
streamlit
opencv-python
numpy
matplotlib
Pillow
scikit-image
plotly
pandas
